🔧 chore: 패키지 구조 변경에 따른 파일 이동
- Event Controller의 payload를 request, response 패키지 하위로 분리
- 아래 파일들을 Event 패키지의 entity 패키지 안으로 이동
- CandidateDate, Event, EventMember, TempSchedule
