+++ src/main/java/com/grepp/spring/app/controller/api/event/payload/response/MyTimeScheduleResponse.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.event.payload;
+package com.grepp.spring.app.controller.api.event.payload.response;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
